# backend/models/user.go
package models

import (
	"time"
)

type User struct {
	ID        uint      `json:"id,omitempty" gorm:"primaryKey;autoIncrement"`
	Name      string    `json:"name"`
	Email     string    `json:"email" gorm:"unique"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}